0-indexed Builder IDs
Inject a phony BInst as id=0 when creating a Builder.  All other code
working with Builders can now assume that every BInst's children exist,
and we can now use simple zero-based indexing into b->inst.

So code like this working with BInst inst's child x,

   if (inst.x && b->inst[inst.x-1]...) ...

now just becomes

   if (b->inst[inst.x]...) ...

We tag the phony id=0 instruction as .kind=CONST so constant_fold() does
not need to check for it explicitly, and as .live=1 so that we don't try
to insert it into our common subexpression hash table, which uses id=0
to indicate an empty slot.