allow load CSE
Track a generation counter for each pointer, bumped when we store to it,
and bake that generation into each load instruction.  This makes stores
invalidate prior loads from that same pointer, including uniform loads.

Each pointer aliasing equivalence class gets its own generation counter.
As written, we assume no pointers alias (A), and so each pointer has its
own counter.  If we allowed aliasing between pointers of the same type
(B), we'd need to track a per-type counter, and if we allowed aliasing
between arbitrary pointers (C), they'd all refer to a single global
generation counter.

   (A) No aliasing:   per-pointer counters, most  optimization
   (B) Strict aliasing:  per-type counters,  ???  optimization
   (C) Relaxed aliasing:   global counter , least optimization

This change implements (A), and (C) would be even easier, and it would
also be easy to implement (A)/(C) as a user option.  (B) would mean we
have to track types, and would involve the most tricky bookkeeping.